+++ src/components/AppWindow/index.tsx
@@ -1,8 +1,9 @@
 import { AppReducerContext } from '@/store/App/AppContext';
-import { RefObject, useContext, useRef, useState } from 'react';
+import { RefObject, useContext } from 'react';
 import CloseIcon from '@/assets/icons/AppWindow/Close.svg?react';
 import MinimizeIcon from '@/assets/icons/AppWindow/Minimize.svg?react';
 import FullScreenIcon from '@/assets/icons/AppWindow/FullScreen.svg?react';
+import useDragElement from '@/hooks/useDragElement';
 
 interface Props {
   title: string;
@@ -14,46 +15,16 @@ const AppWindow = ({ title, id, container }: Props) => {
   const dispatch = useContext(AppReducerContext);
   if (!dispatch) throw new Error('dispatch is null');
 
-  const [{ x, y }, setPosition] = useState({ x: 0, y: 0 });
-  const appWindowRef = useRef<HTMLDivElement>(null);
-
+  const { x, y, targetRef, handleDragElement } = useDragElement<HTMLDivElement>(container, 0, 0);
   const handleClose = () => dispatch({ type: 'CLOSE', id });
-
-  const getPosition = (range: number, min: number, max: number) => {
-    if (range < min) return min;
-    if (range > max) return max;
-    return range;
-  };
-
-  const handleMouseDown = (downEvent: React.MouseEvent) => {
-    const handleMouseMove = (moveEvent: MouseEvent) => {
-      moveEvent.preventDefault();
-      const [moveX, moveY] = [moveEvent.clientX - downEvent.clientX, moveEvent.clientY - downEvent.clientY];
-
-      if (container.current && appWindowRef.current) {
-        const containerRect = container.current.getBoundingClientRect();
-        const appWindowRect = appWindowRef.current.getBoundingClientRect();
-        const calculatedX = getPosition(x + moveX, 0, containerRect.width - appWindowRect.width);
-        const calculatedY = getPosition(y + moveY, 0, containerRect.height - appWindowRect.height);
-        setPosition({ x: calculatedX, y: calculatedY });
-      }
-    };
-
-    const handleMouseUp = () => {
-      document.removeEventListener('mousemove', handleMouseMove);
-    };
-
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp, { once: true });
-  };
 
   return (
     <div
       className='absolute w-96 flex-col overflow-hidden rounded-lg'
       style={{ transform: `translate(${x}px, ${y}px)` }}
-      ref={appWindowRef}
+      ref={targetRef}
     >
-      <div className='flex h-7 w-full items-center justify-center bg-[#e4e4e4]' onMouseDown={handleMouseDown}>
+      <div className='flex h-7 w-full items-center justify-center bg-[#e4e4e4]' onMouseDown={handleDragElement}>
         <div className='hidden-wrapper absolute left-2 flex items-center gap-1'>
           <button
             className='flex h-3 w-3 cursor-pointer items-center justify-center rounded-full border-[0.5px] border-[#00000033] bg-[#FF5F57]'
